+++ ibus-akaza/src/context.rs
@@ -26,8 +26,8 @@ use ibus_sys::glib::gchar;
 use ibus_sys::glib::{gboolean, guint};
 use ibus_sys::lookup_table::{ibus_lookup_table_append_candidate, IBusLookupTable};
 use ibus_sys::text::{ibus_text_new_from_string, ibus_text_set_attributes, StringExt};
-use libakaza::engine::akaza_builder::BigramWordViterbiEngine;
 use libakaza::engine::base::HenkanEngine;
+use libakaza::engine::bigram_word_viterbi_engine::BigramWordViterbiEngine;
 use libakaza::extend_clause::{extend_left, extend_right};
 use libakaza::graph::graph_resolver::Candidate;
 use libakaza::romkan::RomKanConverter;
