+++ libakaza/src/engine/mod.rs
@@ -1,2 +1,2 @@
-pub mod akaza_builder;
 pub mod base;
+pub mod bigram_word_viterbi_engine;
